+++ src/VaultSharp/V1/SecretsEngines/KeyValue/V2/IKeyValueSecretsEngineV2.cs
@@ -116,5 +116,21 @@ namespace VaultSharp.V1.SecretsEngines.KeyValue.V2
         /// The task.
         /// </returns>
         Task DestroySecretAsync(string path, IList<int> versions, string mountPoint = SecretsEngineDefaultPaths.KeyValueV2);
+
+        /// <summary>
+        /// This endpoint permanently deletes the key metadata and all version data for the specified key. 
+        /// All version history will be removed.
+        /// </summary>
+        /// <param name="path"><para>[required]</para>
+        /// Specifies the path of the secret to delete.
+        /// </param>
+        /// <param name="mountPoint"><para>[optional]</para>
+        /// The mount point for the Generic backend. Defaults to <see cref="SecretsEngineDefaultPaths.KeyValueV2" />
+        /// Provide a value only if you have customized the mount point.
+        /// </param>
+        /// <returns>
+        /// The task.
+        /// </returns>
+        Task DeleteMetadataAsync(string path, string mountPoint = SecretsEngineDefaultPaths.KeyValueV2);
     }
 }